push error handling stuff into JsonRepresentation
The YamlRespresentation class doesn't need any of this stuff; push it
into JsonRepresentation to avoid polluting other subclasses of
JacksonRepresentation.
